dnd-web-game/backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlmodel>=0.0.14
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.20.0
alembic>=1.13.0
anthropic>=0.18.0
httpx>=0.26.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pdfplumber>=0.10.0
python-multipart>=0.0.6
